Migrate most utilities out of app
